+++ sdk/resourcemanager/appservice/armappservice/websitemanagement_client.go
@@ -48,7 +48,7 @@ func NewWebSiteManagementClient(subscriptionID string, credential azcore.TokenCr
 // CheckNameAvailability - Description for Check if a resource name is available.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - request - Name availability request.
 //   - options - WebSiteManagementClientCheckNameAvailabilityOptions contains the optional parameters for the WebSiteManagementClient.CheckNameAvailability
 //     method.
@@ -79,7 +79,7 @@ func (client *WebSiteManagementClient) checkNameAvailabilityCreateRequest(ctx co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, request)
@@ -97,7 +97,7 @@ func (client *WebSiteManagementClient) checkNameAvailabilityHandleResponse(resp 
 // GetPublishingUser - Description for Gets publishing user
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientGetPublishingUserOptions contains the optional parameters for the WebSiteManagementClient.GetPublishingUser
 //     method.
 func (client *WebSiteManagementClient) GetPublishingUser(ctx context.Context, options *WebSiteManagementClientGetPublishingUserOptions) (WebSiteManagementClientGetPublishingUserResponse, error) {
@@ -123,7 +123,7 @@ func (client *WebSiteManagementClient) getPublishingUserCreateRequest(ctx contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -141,7 +141,7 @@ func (client *WebSiteManagementClient) getPublishingUserHandleResponse(resp *htt
 // GetSourceControl - Description for Gets source control token
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - sourceControlType - Type of source control
 //   - options - WebSiteManagementClientGetSourceControlOptions contains the optional parameters for the WebSiteManagementClient.GetSourceControl
 //     method.
@@ -172,7 +172,7 @@ func (client *WebSiteManagementClient) getSourceControlCreateRequest(ctx context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -190,7 +190,7 @@ func (client *WebSiteManagementClient) getSourceControlHandleResponse(resp *http
 // GetSubscriptionDeploymentLocations - Description for Gets list of available geo regions plus ministamps
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientGetSubscriptionDeploymentLocationsOptions contains the optional parameters for the WebSiteManagementClient.GetSubscriptionDeploymentLocations
 //     method.
 func (client *WebSiteManagementClient) GetSubscriptionDeploymentLocations(ctx context.Context, options *WebSiteManagementClientGetSubscriptionDeploymentLocationsOptions) (WebSiteManagementClientGetSubscriptionDeploymentLocationsResponse, error) {
@@ -220,7 +220,7 @@ func (client *WebSiteManagementClient) getSubscriptionDeploymentLocationsCreateR
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -235,9 +235,115 @@ func (client *WebSiteManagementClient) getSubscriptionDeploymentLocationsHandleR
 	return result, nil
 }
 
+// GetUsagesInLocation - Gets the usage in cores for a subscription in a given location.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2023-01-01
+//   - location - The location for which to check usages
+//   - options - WebSiteManagementClientGetUsagesInLocationOptions contains the optional parameters for the WebSiteManagementClient.GetUsagesInLocation
+//     method.
+func (client *WebSiteManagementClient) GetUsagesInLocation(ctx context.Context, location string, options *WebSiteManagementClientGetUsagesInLocationOptions) (WebSiteManagementClientGetUsagesInLocationResponse, error) {
+	req, err := client.getUsagesInLocationCreateRequest(ctx, location, options)
+	if err != nil {
+		return WebSiteManagementClientGetUsagesInLocationResponse{}, err
+	}
+	resp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return WebSiteManagementClientGetUsagesInLocationResponse{}, err
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return WebSiteManagementClientGetUsagesInLocationResponse{}, runtime.NewResponseError(resp)
+	}
+	return client.getUsagesInLocationHandleResponse(resp)
+}
+
+// getUsagesInLocationCreateRequest creates the GetUsagesInLocation request.
+func (client *WebSiteManagementClient) getUsagesInLocationCreateRequest(ctx context.Context, location string, options *WebSiteManagementClientGetUsagesInLocationOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Web/locations/{location}/usages"
+	if location == "" {
+		return nil, errors.New("parameter location cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{location}", url.PathEscape(location))
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2023-01-01")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// getUsagesInLocationHandleResponse handles the GetUsagesInLocation response.
+func (client *WebSiteManagementClient) getUsagesInLocationHandleResponse(resp *http.Response) (WebSiteManagementClientGetUsagesInLocationResponse, error) {
+	result := WebSiteManagementClientGetUsagesInLocationResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.CsmUsageQuotaCollection); err != nil {
+		return WebSiteManagementClientGetUsagesInLocationResponse{}, err
+	}
+	return result, nil
+}
+
+// GetZoneRedundantUsagesInLocation - Gets the usage in cores for a subscription in a given location, with zone redundance.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2023-01-01
+//   - location - The location for which to check ZR usages
+//   - options - WebSiteManagementClientGetZoneRedundantUsagesInLocationOptions contains the optional parameters for the WebSiteManagementClient.GetZoneRedundantUsagesInLocation
+//     method.
+func (client *WebSiteManagementClient) GetZoneRedundantUsagesInLocation(ctx context.Context, location string, options *WebSiteManagementClientGetZoneRedundantUsagesInLocationOptions) (WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse, error) {
+	req, err := client.getZoneRedundantUsagesInLocationCreateRequest(ctx, location, options)
+	if err != nil {
+		return WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse{}, err
+	}
+	resp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse{}, err
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse{}, runtime.NewResponseError(resp)
+	}
+	return client.getZoneRedundantUsagesInLocationHandleResponse(resp)
+}
+
+// getZoneRedundantUsagesInLocationCreateRequest creates the GetZoneRedundantUsagesInLocation request.
+func (client *WebSiteManagementClient) getZoneRedundantUsagesInLocationCreateRequest(ctx context.Context, location string, options *WebSiteManagementClientGetZoneRedundantUsagesInLocationOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Web/locations/{location}/zoneRedundant/usages"
+	if location == "" {
+		return nil, errors.New("parameter location cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{location}", url.PathEscape(location))
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2023-01-01")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// getZoneRedundantUsagesInLocationHandleResponse handles the GetZoneRedundantUsagesInLocation response.
+func (client *WebSiteManagementClient) getZoneRedundantUsagesInLocationHandleResponse(resp *http.Response) (WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse, error) {
+	result := WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.CsmUsageQuotaCollection); err != nil {
+		return WebSiteManagementClientGetZoneRedundantUsagesInLocationResponse{}, err
+	}
+	return result, nil
+}
+
 // NewListBillingMetersPager - Description for Gets a list of meters for a given location.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientListBillingMetersOptions contains the optional parameters for the WebSiteManagementClient.NewListBillingMetersPager
 //     method.
 func (client *WebSiteManagementClient) NewListBillingMetersPager(options *WebSiteManagementClientListBillingMetersOptions) *runtime.Pager[WebSiteManagementClientListBillingMetersResponse] {
@@ -286,7 +392,7 @@ func (client *WebSiteManagementClient) listBillingMetersCreateRequest(ctx contex
 	if options != nil && options.OSType != nil {
 		reqQP.Set("osType", *options.OSType)
 	}
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -303,7 +409,7 @@ func (client *WebSiteManagementClient) listBillingMetersHandleResponse(resp *htt
 
 // NewListCustomHostNameSitesPager - Get custom hostnames under this subscription
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientListCustomHostNameSitesOptions contains the optional parameters for the WebSiteManagementClient.NewListCustomHostNameSitesPager
 //     method.
 func (client *WebSiteManagementClient) NewListCustomHostNameSitesPager(options *WebSiteManagementClientListCustomHostNameSitesOptions) *runtime.Pager[WebSiteManagementClientListCustomHostNameSitesResponse] {
@@ -349,7 +455,7 @@ func (client *WebSiteManagementClient) listCustomHostNameSitesCreateRequest(ctx 
 	if options != nil && options.Hostname != nil {
 		reqQP.Set("hostname", *options.Hostname)
 	}
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -366,7 +472,7 @@ func (client *WebSiteManagementClient) listCustomHostNameSitesHandleResponse(res
 
 // NewListGeoRegionsPager - Description for Get a list of available geographical regions.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientListGeoRegionsOptions contains the optional parameters for the WebSiteManagementClient.NewListGeoRegionsPager
 //     method.
 func (client *WebSiteManagementClient) NewListGeoRegionsPager(options *WebSiteManagementClientListGeoRegionsOptions) *runtime.Pager[WebSiteManagementClientListGeoRegionsResponse] {
@@ -421,7 +527,7 @@ func (client *WebSiteManagementClient) listGeoRegionsCreateRequest(ctx context.C
 	if options != nil && options.LinuxDynamicWorkersEnabled != nil {
 		reqQP.Set("linuxDynamicWorkersEnabled", strconv.FormatBool(*options.LinuxDynamicWorkersEnabled))
 	}
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -438,7 +544,7 @@ func (client *WebSiteManagementClient) listGeoRegionsHandleResponse(resp *http.R
 
 // NewListPremierAddOnOffersPager - Description for List all premier add-on offers.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientListPremierAddOnOffersOptions contains the optional parameters for the WebSiteManagementClient.NewListPremierAddOnOffersPager
 //     method.
 func (client *WebSiteManagementClient) NewListPremierAddOnOffersPager(options *WebSiteManagementClientListPremierAddOnOffersOptions) *runtime.Pager[WebSiteManagementClientListPremierAddOnOffersResponse] {
@@ -481,7 +587,7 @@ func (client *WebSiteManagementClient) listPremierAddOnOffersCreateRequest(ctx c
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -499,7 +605,7 @@ func (client *WebSiteManagementClient) listPremierAddOnOffersHandleResponse(resp
 // ListSKUs - Description for List all SKUs.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientListSKUsOptions contains the optional parameters for the WebSiteManagementClient.ListSKUs
 //     method.
 func (client *WebSiteManagementClient) ListSKUs(ctx context.Context, options *WebSiteManagementClientListSKUsOptions) (WebSiteManagementClientListSKUsResponse, error) {
@@ -529,7 +635,7 @@ func (client *WebSiteManagementClient) listSKUsCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -546,7 +652,7 @@ func (client *WebSiteManagementClient) listSKUsHandleResponse(resp *http.Respons
 
 // NewListSiteIdentifiersAssignedToHostNamePager - Description for List all apps that are assigned to a hostname.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - nameIdentifier - Hostname information.
 //   - options - WebSiteManagementClientListSiteIdentifiersAssignedToHostNameOptions contains the optional parameters for the
 //     WebSiteManagementClient.NewListSiteIdentifiersAssignedToHostNamePager method.
@@ -590,7 +696,7 @@ func (client *WebSiteManagementClient) listSiteIdentifiersAssignedToHostNameCrea
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, nameIdentifier)
@@ -607,7 +713,7 @@ func (client *WebSiteManagementClient) listSiteIdentifiersAssignedToHostNameHand
 
 // NewListSourceControlsPager - Description for Gets the source controls available for Azure websites.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - options - WebSiteManagementClientListSourceControlsOptions contains the optional parameters for the WebSiteManagementClient.NewListSourceControlsPager
 //     method.
 func (client *WebSiteManagementClient) NewListSourceControlsPager(options *WebSiteManagementClientListSourceControlsOptions) *runtime.Pager[WebSiteManagementClientListSourceControlsResponse] {
@@ -646,7 +752,7 @@ func (client *WebSiteManagementClient) listSourceControlsCreateRequest(ctx conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -664,7 +770,7 @@ func (client *WebSiteManagementClient) listSourceControlsHandleResponse(resp *ht
 // Move - Description for Move resources between resource groups.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - resourceGroupName - Name of the resource group to which the resource belongs.
 //   - moveResourceEnvelope - Object that represents the resource to move.
 //   - options - WebSiteManagementClientMoveOptions contains the optional parameters for the WebSiteManagementClient.Move method.
@@ -699,7 +805,7 @@ func (client *WebSiteManagementClient) moveCreateRequest(ctx context.Context, re
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, moveResourceEnvelope)
@@ -708,7 +814,7 @@ func (client *WebSiteManagementClient) moveCreateRequest(ctx context.Context, re
 // UpdatePublishingUser - Description for Updates publishing user
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - userDetails - Details of publishing user
 //   - options - WebSiteManagementClientUpdatePublishingUserOptions contains the optional parameters for the WebSiteManagementClient.UpdatePublishingUser
 //     method.
@@ -735,7 +841,7 @@ func (client *WebSiteManagementClient) updatePublishingUserCreateRequest(ctx con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, userDetails)
@@ -753,7 +859,7 @@ func (client *WebSiteManagementClient) updatePublishingUserHandleResponse(resp *
 // UpdateSourceControl - Description for Updates source control token
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - sourceControlType - Type of source control
 //   - requestMessage - Source control token information
 //   - options - WebSiteManagementClientUpdateSourceControlOptions contains the optional parameters for the WebSiteManagementClient.UpdateSourceControl
@@ -785,7 +891,7 @@ func (client *WebSiteManagementClient) updateSourceControlCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, requestMessage)
@@ -803,7 +909,7 @@ func (client *WebSiteManagementClient) updateSourceControlHandleResponse(resp *h
 // Validate - Description for Validate if a resource can be created.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - resourceGroupName - Name of the resource group to which the resource belongs.
 //   - validateRequest - Request with the resources to validate.
 //   - options - WebSiteManagementClientValidateOptions contains the optional parameters for the WebSiteManagementClient.Validate
@@ -839,7 +945,7 @@ func (client *WebSiteManagementClient) validateCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, validateRequest)
@@ -857,7 +963,7 @@ func (client *WebSiteManagementClient) validateHandleResponse(resp *http.Respons
 // ValidateMove - Description for Validate whether a resource can be moved.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - resourceGroupName - Name of the resource group to which the resource belongs.
 //   - moveResourceEnvelope - Object that represents the resource to move.
 //   - options - WebSiteManagementClientValidateMoveOptions contains the optional parameters for the WebSiteManagementClient.ValidateMove
@@ -893,7 +999,7 @@ func (client *WebSiteManagementClient) validateMoveCreateRequest(ctx context.Con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, moveResourceEnvelope)
@@ -903,7 +1009,7 @@ func (client *WebSiteManagementClient) validateMoveCreateRequest(ctx context.Con
 // the Network Security Group rules.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-01
+// Generated from API version 2023-01-01
 //   - parameters - VNET information
 //   - options - WebSiteManagementClientVerifyHostingEnvironmentVnetOptions contains the optional parameters for the WebSiteManagementClient.VerifyHostingEnvironmentVnet
 //     method.
@@ -934,7 +1040,7 @@ func (client *WebSiteManagementClient) verifyHostingEnvironmentVnetCreateRequest
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-01")
+	reqQP.Set("api-version", "2023-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, parameters)